+++ src/app/components/product-card/product-card.component.html
@@ -10,10 +10,4 @@
     </div>
 
   </div>
-
-  <div class="card-body row ">
-    <!-- <button class="btn btn-outline-success mb-1" id="button-style" (click)="addToCart(productInfo)">Add to Cart</button>
-    <button class="btn btn-outline-success mt-1" id="button-style" (click)="addToWishCart(productInfo)">Add to Wish
-      List</button> -->
-  </div>
 </div>
