+++ src/app/components/product-details/product-details.component.css
@@ -1,0 +1,4 @@
+.card-image{
+    width:30rem;
+
+}
